+++ CoineyKit/CoineyKit.framework/Headers/CYContainerViewController.h
@@ -5,11 +5,12 @@
 #import <UIKit/UIKit.h>
 
 @interface CYContainerViewController : UIViewController
-@property(nonatomic, strong) UIViewController *containedViewController;
-- (void)setContainedViewController:(UIViewController *)aController animated:(BOOL)aAnimated;
-- (void)setContainedViewController:(UIViewController *)aController
+@property(nonatomic, strong, nullable) UIViewController *containedViewController;
+- (void)setContainedViewController:(UIViewController * _Nonnull)aController
+                          animated:(BOOL)aAnimated;
+- (void)setContainedViewController:(UIViewController * _Nonnull)aController
                           animated:(BOOL)aAnimated
-                        completion:(void (^)(void))aBlock;
+                        completion:(void (^_Nullable)(void))aBlock;
 @end
 
 @interface CYContainmentSegue : UIStoryboardSegue